inicio das correções

+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: pvitor-l <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/16 15:50:49 by pvitor-l          #+#    #+#             */
-/*   Updated: 2024/10/16 19:55:05 by pvitor-l         ###   ########.fr       */
+/*   Updated: 2024/10/22 14:52:33 by pvitor-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*copy;
 	unsigned char	*source;
